Move TOC component to public/
This moves the TOC React component to the `public/` dir since it's only
used in the browser, and updates the build to handle JSX in the
`public/` dir.
