+++ public/js/components/table-of-contents.jsx
